back-end/requirements.txt
appnope==0.1.0
argcomplete==1.11.0
asgiref==3.2.5
asn1crypto==1.3.0
atomicwrites==1.3.0
attrs==19.3.0
awscli==1.16.298
backcall==0.1.0
bamread==0.0.5
boto3==1.10.39
botocore==1.13.34
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.1
cryptography==2.8
Cython==0.29.14
decorator==4.4.2
Django==3.0.3
django-cors-headers==3.0.2
django-environ==0.4.5
django-filter==2.2.0
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
fsspec==0.7.1
funcsigs==1.0.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.9.4
jsonschema==3.1.1
jupyter-client==6.1.3
jupyter-core==4.6.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
natsort==7.0.1
ncls==0.0.50
numpy==1.18.1
packaging==19.2
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==3.0.5
protobuf==3.10.0
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pybedtools==0.8.0
pyBigWig==0.3.17
pycparser==2.20
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyranges==0.0.62
pyrle==0.0.25
pyrsistent==0.15.5
pysam==0.15.3
PySocks==1.7.1
pytest==5.2.2
python-dateutil==2.8.0
python-dotenv==0.12.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
ray==0.7.6
redis==3.3.11
requests==2.22.0
rsa==3.4.2
s3fs==0.4.0
s3transfer==0.2.1
scipy==1.4.1
sh==1.12.14
six==1.14.0
sorted-nearest==0.0.30
sqlparse==0.3.0
tabulate==0.8.3
tornado==6.0.4
traitlets==4.3.3
urllib3==1.24.3
wcwidth==0.1.9
xmltodict==0.12.0
yq==2.10.0
zipp==0.6.0
